docstring, file renaming, accumulator changes
